abac_poc/requirements.txt
Django==2.2.7
django-rest-framework-mongoengine==3.4.0
djangorestframework==3.10.3
mongoengine==0.18.2
requests==2.20.0
